feat: add integration tests for index and segmentation functionalities
