+++ tests/integration/conftest.py
@@ -1,5 +1,4 @@
 # tests/integration/conftest.py
-
 import pytest
 from pathlib import Path
