+++ tests/integration/coretypes/masktypes/seg.py
