+++ tests/integration/coretypes/masktypes/structureset.py
